requirement.txt
asttokens==2.1.0
backcall==0.2.0
click==8.1.3
colorama==0.4.5
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
face-recognition-models==0.3.0
Flask==2.1.2
importlib-metadata==4.11.4
ipykernel==6.17.1
ipython==8.6.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter_client==7.4.5
jupyter_core==5.0.0
jupyterlab-widgets==1.1.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pickleshare==0.7.5
platformdirs==2.5.4
prompt-toolkit==3.0.32
psutil==5.9.4
pure-eval==0.2.2
pybuilder==0.13.8
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
pywin32==304
pyzmq==24.0.1
scipy==1.8.1
six==1.16.0
stack-data==0.6.1
tornado==6.2
traitlets==5.5.0
wcwidth==0.2.5
Werkzeug==2.1.2
xgboost==1.6.1
zipp==3.8.0
